duplicate_detection_tool/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
urllib3==2.1.0
google-generativeai==0.3.2
chromadb==0.4.22
Pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6

